Bugfix in users
