+++ api_yamdb/api/urls.py
@@ -2,7 +2,10 @@ from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 
 from .views import (CategoryViewSet, CommentViewSet, GenreViewSet,
-                    ReviewViewSet, TitleViewSet)
+                    ReviewViewSet, TitleViewSet, UserViewSet, get_jwt_token,
+                    signup)
+
+app_name = 'users'
 
 router_v1 = DefaultRouter()
 router_v1.register(
@@ -18,7 +21,12 @@ router_v1.register(
 router_v1.register('categories', CategoryViewSet)
 router_v1.register('genres', GenreViewSet)
 router_v1.register('titles', TitleViewSet)
+router_v1.register(r'users', UserViewSet, basename='users')
+
 
 urlpatterns = [
+    path('v1/', include(router_v1.urls)),
+    path('v1/auth/token/', get_jwt_token, name='token'),
+    path('v1/auth/signup/', signup),
     path('v1/', include(router_v1.urls)),
 ]
